+++ src/main/java/com/ganesh/service/EmployeeServiceImpl.java
@@ -26,7 +26,7 @@ public class EmployeeServiceImpl implements EmployeeService {
     }
 
     @Override
-    public Employee getEmployeeById(int id) {
+    public Employee getEmployeeById(long id) {
         return empDao.getEmployeeById(id);
     }
 
@@ -36,7 +36,7 @@ public class EmployeeServiceImpl implements EmployeeService {
     }
 
     @Override
-    public boolean deleteEmployee(int id) {
+    public boolean deleteEmployee(long id) {
         return empDao.deleteEmployee(id);
     }
